requirements.txt
# Default Python packages:
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4

# sqlite
SQLAlchemy~=2.0.1

# ssh
paramiko~=3.0.0
#
pyyaml~=6.0

pytest~=7.2.1
# termcolor~=2.2.0
string-color==1.2.3
